Rename files for consistency
