+++ src/query/SetDayQuery.rb
