+++ src/query/SetTimeQuery.rb
